Reafactor: All Utils made to JSON
