+++ go.mod
@@ -4,10 +4,10 @@ go 1.16
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/dustin/go-humanize v1.0.0
 	github.com/gocarina/gocsv v0.0.0-20210408192840-02d7211d929d
 	github.com/jmoiron/sqlx v1.3.3
 	github.com/labstack/echo/v4 v4.2.2
 	github.com/lib/pq v1.10.0
 	github.com/stretchr/testify v1.6.0 // indirect
+	golang.org/x/tools v0.1.3 // indirect
 )
