--- templates/index.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Kanakkapilla Home</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@exampledev/new.css@1.1.2/new.min.css">
-    <script src="https://unpkg.com/frappe-charts@1.2.4/dist/frappe-charts.min.iife.js"></script>
-
-</head>
-<body>
-    <nav>
-        <a href="/all">All Transactions</a>
-        <a href="/upi">UPI Transactions</a>
-        <form action="/search" method="POST">
-            Type What you want to search for: <input type="search" name="query">
-            <input type="submit"/>
-            </form>
-    </nav>
-<h3>Current Bank ,Balance Amount as on {{.BalonDate | hdate}}: <mark>{{.CurBal | humanize | currit}}</mark></h3>
-    <hr>
-    <h3>Maximum Debited Amount from UPI: {{.UPIPoints.MxDebit | humanize | currit}}</h3>
-    <h3>Maximum Credit Amount from UPI: {{.UPIPoints.MxCredit | humanize | currit}}</h3>
-    <hr/>
-    <h3>Sum of Transactions in UPI</h3>
-    <h4>Total Debited Amount:{{.UPISum.DebSum| humanize | currit}}</h4>
-    <h4>Total Credited Amount: {{.UPISum.CredSum| humanize | currit}}</h4>
-    <hr/>
-    <hr/>
-    <div id="chart-1"></div>
-    <hr>
-    <div id="chart-2"></div>
-</body>
-<style>
-    body{
-        max-width:max-content !important;
-        }
-</style>
-</html>
