--- templates/search-table.html
@@ -1,44 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Table</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@exampledev/new.css@1.1.2/new.min.css">
-
-</head>
-<body>
-    <nav>
-        <a href="/">UPI Transactions</a>
-    </nav>
-    <table>
-        <tr>
-            <th>ID</th>
-            <th>TDate</th>
-            <th>Date</th>
-            <th>Description</th>
-            <th>Ref</th>
-            <th>Debit</th>
-            <th>Credit</th>
-            <th>Balance</th>
-        </tr>
-        {{ range .}}
-            <tr>
-                <td>{{ .ID }}</td>
-                <td>{{ .Tdate|hdate }}</td>
-                <td>{{ .Date|hdate }}</td>
-                <td>{{ .Description }}</td>
-                <td>{{ .Refno.String }}</td>
-                <td>{{ .Debit.Float64| humanize | currit }}</td>
-                <td>{{ .Credit.Float64| humanize | currit }}</td>
-                <td>{{ .Balance| humanize | currit }}</td>
-            </tr>
-        {{ end}}
-    </table>
-</body>
-<style>
-    body{
-        max-width:max-content !important;
-        }
-</style>
-</html>
