--- templates/upi.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Table</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@exampledev/new.css@1.1.2/new.min.css">
-
-</head>
-<body>
-    <nav>
-        <a href="/">Home</a>
-        <a href="/all">All Transactions</a>
-    </nav>
-<ol>
-    <hr>
-    {{- range .UPITrans }}
-    <li><mark>{{.Date|hdate}}</mark></li>
-    <ul>
-    {{- if not .Credit.Valid}}
-    <li>{{.Description}}</li>
-    <li>{{.Refno.String}}</li>
-    <li>Debited: 🙈 {{ .Debit.Float64 | humanize | currit }}</li>
-    {{- else}}
-    <li>{{.Description}}</li>
-    <li>{{.Refno.String}}</li>
-    <li>Credit: 🤑 {{ .Credit.Float64 | humanize | currit}}</li>
-    {{end -}}
-    </ul>
-    
-    {{- end}}
-    </ol>
-</body>
-</html>
